rag-code/requirements.txt
google-generativeai
python-dotenv
cohere
qdrant-client
trafilatura
requests
lxml
openai-agents
fastapi
uvicorn
tinyagents
sqlean.py